+++ ChangeLog.md
@@ -9,6 +9,7 @@
 - Fixed issue that x-ms-blob-sequence-number of 0 should be returned for HEAD requests on Page blob.
 - Uploading blocks with different lengths of IDs to the same blob will fail.
 - Check if block blob exists should fail if blocks are all uncommitted.
+- Case sensitive with metadata keys.
 
 ## 2020.02 Version 3.5.0